// src/client/types.gen.ts
export type CreateExternalApplicationModel = {
    /**
     * ФИО заказчика
     */
    fullName: string;
    /**
     * Адрес почты заказчика
     */
    email: string;
    /**
     * Номер телефона заказчика
     */
    phone?: string | null;
    /**
     * Название организации
     */
    organisationName: string;
    /**
     * Ссылка на сайт организации
     */
    organisationUrl?: string | null;
    /**
     * Название проекта
     */
    projectName: string;
    /**
     * Id типа проекта
     */
    typeId: number;
    /**
     * Ожидаемые результаты
     */
    expectedResults: string;
    /**
     * Предполагается ли оплата
     */
    isPayed: boolean;
    /**
     * Дополнительная информация
     */
    additionalInformation?: string | null;
};